Inherited NFAStateGraph and NFAStateGraph builder from package-private abstract class to eliminate code duplication.
